+++ CHANGELOG.md
@@ -243,3 +243,4 @@
 * Improved a few edge cases in ActorFollowerTetherEx
 * PlayMusic no longer restarts songs if they are already playing
 * Fixed ThemeableElement and NonThemeableElement being ignored on child components of complex elements [Untested]
+* Fixed dialogue reactivating hidden objects even if they were initially inactive [Untested]